[minor] Split stats.h into one header per class
